+++ src/wavtools.rs
@@ -4,7 +4,7 @@ use std::io::BufReader;
 use std::io::{Read, Seek, SeekFrom, Write};
 use std::mem;
 
-use crate::config::{ConfigError, FileFormat, SampleFormat};
+use crate::config::{ConfigError, SampleFormat};
 use crate::Res;
 
 const RIFF: &[u8] = "RIFF".as_bytes();
@@ -55,7 +55,7 @@ const SUBTYPE_PCM: Guid = Guid {
 
 #[derive(Debug)]
 pub struct WavParams {
-    pub sample_format: FileFormat,
+    pub sample_format: SampleFormat,
     pub sample_rate: usize,
     pub data_offset: usize,
     pub data_length: usize,
@@ -88,14 +88,14 @@ fn look_up_format(
     bits: u16,
     bytes_per_sample: u16,
     chunk_length: u32,
-) -> Res<FileFormat> {
+) -> Res<SampleFormat> {
     match (formatcode, bits, bytes_per_sample) {
-        (1, 16, 2) => Ok(FileFormat::S16LE),
-        (1, 24, 3) => Ok(FileFormat::S24LE3),
-        (1, 24, 4) => Ok(FileFormat::S24LE),
-        (1, 32, 4) => Ok(FileFormat::S32LE),
-        (3, 32, 4) => Ok(FileFormat::FLOAT32LE),
-        (3, 64, 8) => Ok(FileFormat::FLOAT64LE),
+        (1, 16, 2) => Ok(SampleFormat::S16LE),
+        (1, 24, 3) => Ok(SampleFormat::S24LE3),
+        (1, 24, 4) => Ok(SampleFormat::S24LE),
+        (1, 32, 4) => Ok(SampleFormat::S32LE),
+        (3, 32, 4) => Ok(SampleFormat::FLOAT32LE),
+        (3, 64, 8) => Ok(SampleFormat::FLOAT64LE),
         (0xFFFE, _, _) => look_up_extended_format(data, bits, bytes_per_sample, chunk_length),
         (_, _, _) => Err(ConfigError::new("Unsupported wav format").into()),
     }
@@ -106,7 +106,7 @@ fn look_up_extended_format(
     bits: u16,
     bytes_per_sample: u16,
     chunk_length: u32,
-) -> Res<FileFormat> {
+) -> Res<SampleFormat> {
     if chunk_length != 40 {
         return Err(ConfigError::new("Invalid extended header").into());
     }
@@ -125,12 +125,12 @@ fn look_up_extended_format(
         bytes_per_sample,
         valid_bits_per_sample,
     ) {
-        (SUBTYPE_PCM, 16, 2, 16) => Ok(FileFormat::S16LE),
-        (SUBTYPE_PCM, 24, 3, 24) => Ok(FileFormat::S24LE3),
-        (SUBTYPE_PCM, 24, 4, 24) => Ok(FileFormat::S24LE),
-        (SUBTYPE_PCM, 32, 4, 32) => Ok(FileFormat::S32LE),
-        (SUBTYPE_FLOAT, 32, 4, 32) => Ok(FileFormat::FLOAT32LE),
-        (SUBTYPE_FLOAT, 64, 8, 64) => Ok(FileFormat::FLOAT64LE),
+        (SUBTYPE_PCM, 16, 2, 16) => Ok(SampleFormat::S16LE),
+        (SUBTYPE_PCM, 24, 3, 24) => Ok(SampleFormat::S24LE3),
+        (SUBTYPE_PCM, 24, 4, 24) => Ok(SampleFormat::S24LE),
+        (SUBTYPE_PCM, 32, 4, 32) => Ok(SampleFormat::S32LE),
+        (SUBTYPE_FLOAT, 32, 4, 32) => Ok(SampleFormat::FLOAT32LE),
+        (SUBTYPE_FLOAT, 64, 8, 64) => Ok(SampleFormat::FLOAT64LE),
         (_, _, _, _) => Err(ConfigError::new("Unsupported extended wav format").into()),
     }
 }
@@ -158,7 +158,7 @@ pub fn find_data_in_wav_stream(mut f: impl Read + Seek) -> Res<WavParams> {
     // Bytes 8 to 12 must be WAVE
     let wave_err = !compare_4cc(&header[8..], WAVE);
     if riff_err || wave_err {
-        return Err(ConfigError::new("Invalid header").into());
+        return Err(ConfigError::new("Invalid wav header").into());
     }
 
     let mut next_chunk_location = 12;
@@ -167,7 +167,7 @@ pub fn find_data_in_wav_stream(mut f: impl Read + Seek) -> Res<WavParams> {
     let mut buffer = [0; 8];
 
     // Dummy values until we have found the real ones
-    let mut sample_format = FileFormat::S16LE;
+    let mut sample_format = SampleFormat::S16LE;
     let mut sample_rate = 0;
     let mut channels = 0;
     let mut data_offset = 0;
@@ -276,14 +276,14 @@ mod tests {
     use super::find_data_in_wav;
     use super::find_data_in_wav_stream;
     use super::write_wav_header;
-    use crate::config::{FileFormat, SampleFormat};
+    use crate::config::SampleFormat;
     use std::io::Cursor;
 
     #[test]
     pub fn test_analyze_wav() {
         let info = find_data_in_wav("testdata/int32.wav").unwrap();
         println!("{info:?}");
-        assert_eq!(info.sample_format, FileFormat::S32LE);
+        assert_eq!(info.sample_format, SampleFormat::S32LE);
         assert_eq!(info.data_offset, 44);
         assert_eq!(info.data_length, 20);
         assert_eq!(info.channels, 1);
@@ -294,7 +294,7 @@ mod tests {
     pub fn test_analyze_wavex() {
         let info = find_data_in_wav("testdata/f32_ex.wav").unwrap();
         println!("{info:?}");
-        assert_eq!(info.sample_format, FileFormat::FLOAT32LE);
+        assert_eq!(info.sample_format, SampleFormat::FLOAT32LE);
         assert_eq!(info.data_offset, 104);
         assert_eq!(info.data_length, 20);
         assert_eq!(info.channels, 1);
@@ -307,7 +307,7 @@ mod tests {
         let mut buffer = Cursor::new(bytes);
         write_wav_header(&mut buffer, 2, SampleFormat::S32LE, 44100).unwrap();
         let info = find_data_in_wav_stream(buffer).unwrap();
-        assert_eq!(info.sample_format, FileFormat::S32LE);
+        assert_eq!(info.sample_format, SampleFormat::S32LE);
         assert_eq!(info.data_offset, 44);
         assert_eq!(info.channels, 2);
         assert_eq!(info.sample_rate, 44100);